+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Configurable simplification steps (#54)
 - Parallel simplification (#53)
 - Terminate fuzzer workers in atexit() handler
+- Subcommand to prune invalid crashes (#52)
 
 ### Changed